+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Support for having conditions
 - Literal condition values for first party mode
+- Support for morph relations
 
 ## [0.2.1] - 2023-07-03
 ### Added